+++ backend/src/Backend/Import.hs
@@ -9,23 +9,21 @@
 {-# LANGUAGE TypeFamilies #-}
 {-# LANGUAGE TypeSynonymInstances #-}
 {-# LANGUAGE UndecidableInstances #-}
+
 -- Import Slack JSON archive into database
 module Backend.Import where
 
-import Control.Lens.Operators ((<&>))
 import Control.Monad
 import Data.Aeson
 import qualified Data.ByteString.Lazy as BL
-import Data.List (isSuffixOf)
 import Data.List.Split (chunksOf)
 import Data.Map (Map)
 import qualified Data.Map as Map
 import Data.Maybe (listToMaybe)
 import Data.Text (Text)
 import qualified Data.Text as T
-import System.Directory
-import System.FilePath ((</>))
 
+import Codec.Archive.Zip
 import Database.Beam
 import Database.Beam.Sqlite (Sqlite, runBeamSqlite)
 import qualified Database.SQLite.Simple as SQLite
@@ -33,8 +31,9 @@ import qualified Database.SQLite.Simple as SQLite
 import Common.Slack.Types
 import Common.Slack.Types.Auth (SlackTeam (..))
 
-rootDir :: String
-rootDir = "/home/srid/code/Taut/tmp"
+-- TODO: This should come from config file
+archivePath :: FilePath
+archivePath = "/home/srid/Downloads/ActualFreedom Slack export Jul 10 2016 - May 10 2019.zip"
 
 data SlackDb f = SlackDb
   { _slackChannels :: f (TableEntity ChannelT)
@@ -48,36 +47,37 @@ instance Database be SlackDb
 slackDb :: DatabaseSettings be SlackDb
 slackDb = defaultDbSettings
 
-loadFile :: FromJSON a => String -> IO a
-loadFile = BL.readFile >=> either error return . eitherDecode
+loadFromArchive :: FromJSON a => EntrySelector -> ZipArchive a
+loadFromArchive = either error pure . eitherDecode <=< fmap BL.fromStrict . getEntry
 
-channelMessages :: Channel -> IO [Message]
-channelMessages channel = msgFiles >>= fmap join . traverse loadFile >>= pure . fmap setChannel
+channelMessages :: Channel -> ZipArchive [Message]
+channelMessages channel = fmap setChannel <$> (msgFiles >>= fmap join . traverse loadFromArchive)
   where
     msgFiles = do
-      let c = T.unpack $ _channelName channel
-      listDirectory (rootDir </> c)
-        <&> filter (isSuffixOf ".json")
-        <&> fmap ((rootDir </> c) </>)
+      let c = _channelName channel
+      entries <- withArchive archivePath (Map.keys <$> getEntries)
+      pure $ flip filter entries $ \e ->
+        let name = getEntryName e in T.isPrefixOf (c <> "/") name && T.isSuffixOf ".json" name
     setChannel msg = msg  { _messageChannelName = Just $ _channelName channel }
 
--- TODO: Figure out a better way to do this. beam-migrate?
 schema :: [SQLite.Query]
 schema =
-  [ "CREATE TABLE channels (id VARCHAR NOT NULL, name VARCHAR NOT NULL, created VARCHAR NOT NULL, PRIMARY KEY( id ));"
+  [ "CREATE TABLE channels (id VARCHAR NOT NULL, name VARCHAR NOT NULL, PRIMARY KEY( id ));"
   , "CREATE TABLE users (id VARCHAR NOT NULL, team_id VARCHAR NOT NULL, name VARCHAR NOT NULL, deleted BOOL NOT NULL, color VARCHAR, real_name VARCHAR, tz VARCHAR, tz_label VARCHAR, tz_offset INTEGER, PRIMARY KEY( id ));"
   , "CREATE TABLE messages (id INTEGER PRIMARY KEY, type VARCHAR NOT NULL, subtype VARCHAR, user VARCHAR, user_name VARCHAR, bot_id VARCHAR, text VARCHAR NOT NULL, client_msg_id VARCHAR, ts INT NOT NULL, channel_name VARCHAR);"
   ]
 
 populateDatabase :: SQLite.Connection -> IO SlackTeam
 populateDatabase conn = do
-  users <- loadFile $ rootDir </> "users.json" :: IO [User]
-  channels <- loadFile $ rootDir </> "channels.json" :: IO [Channel]
+  (users, channels, messages) <- withArchive archivePath $ do
+    users :: [User] <- loadFromArchive =<< mkEntrySelector "users.json"
+    channels :: [Channel] <- loadFromArchive =<< mkEntrySelector "channels.json"
+    messages <- fmap join $ traverse channelMessages channels
+    pure (users, channels, messages)
+
   let Just (team :: SlackTeam) = fmap SlackTeam $ listToMaybe $ _userTeamId <$> users
-
-  messages <- fmap join $ traverse channelMessages channels
-
   let userMap = Map.fromList $ flip fmap users $ \u -> (_userId u, _userName u)
+
   SQLite.withTransaction conn $ do
     -- Create tables
     forM_ schema $ SQLite.execute_ conn